@@ -4,9 +4,9 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.aldebaran.utils.gone
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.themovie.databinding.BottomSheetNetworkBinding
-import com.themovie.helper.gone
 
 class NoInternetBottomSheet(private val hideRetry: Boolean): BottomSheetDialogFragment() {
 
