@@ -2,12 +2,14 @@ package com.themovie.ui.detail
 
 import android.os.Bundle
 import androidx.navigation.findNavController
-import com.themovie.R
-import com.themovie.base.BaseActivity
-import com.themovie.databinding.ActivityDetailBinding
-import com.themovie.di.detail.DetailComponent
-import com.themovie.helper.Constant
+import com.aldebaran.base.BaseActivity
 
+import com.themovie.R
+import com.themovie.databinding.ActivityDetailBinding
+import com.themovie.helper.Constant
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
 class DetailActivity : BaseActivity<ActivityDetailBinding>() {
 
     override fun getLayout(): Int {
@@ -29,9 +31,5 @@ class DetailActivity : BaseActivity<ActivityDetailBinding>() {
     override fun onSupportNavigateUp(): Boolean {
         super.onBackPressed()
         return true
-    }
-
-    fun getDetailComponent(): DetailComponent {
-        return getApp().getAppComponent().detailMovieComponent().create()
     }
 }
