@@ -4,42 +4,36 @@ package com.themovie.ui.detail
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.View
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import androidx.navigation.Navigation
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.aldebaran.domain.Result.Status.*
+import com.aldebaran.base.BaseFragment
+import com.aldebaran.domain.entities.remote.*
+import com.aldebaran.utils.changeActivity
+import com.aldebaran.utils.gone
+import com.aldebaran.utils.navigateFragment
+import com.aldebaran.utils.visible
 
 import com.themovie.R
-import com.themovie.base.BaseFragment
 import com.themovie.databinding.FragmentDetailTvBinding
-import com.themovie.di.detail.DetailViewModelFactory
-import com.themovie.helper.*
-import com.themovie.model.online.detail.Credits
-import com.themovie.model.online.detail.Reviews
-import com.themovie.model.db.Tv
-import com.themovie.model.online.video.Videos
-import com.themovie.restapi.Result
-import com.themovie.restapi.Result.Status.SUCCESS
+import com.themovie.helper.Constant
+import com.themovie.helper.showNetworkError
+
 import com.themovie.ui.detail.adapter.*
 import com.themovie.ui.detail.viewmodel.DetailTvViewModel
-import javax.inject.Inject
+import com.themovie.ui.youtube.YoutubeActivity
+import dagger.hilt.android.AndroidEntryPoint
 
-
-/**
- * A simple [Fragment] subclass.
- *
- */
+@AndroidEntryPoint
 class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
 
-    @Inject lateinit var factory: DetailViewModelFactory
-    private val viewModel by viewModels<DetailTvViewModel> { factory }
-    private lateinit var seasonAdapter: SeasonAdapter
-    private lateinit var creditsAdapter: CreditsAdapter
-    private lateinit var recommendedTvAdapter: RecommendedTvAdapter
-    private lateinit var reviewsAdapter: ReviewsAdapter
-    private lateinit var videoAdapter: VideoAdapter
+    private val viewModel by viewModels<DetailTvViewModel>()
+    private val seasonAdapter by lazy { SeasonAdapter() }
+    private val creditsAdapter by lazy { CreditsAdapter(::onClickCreditItem) }
+    private val recommendedTvAdapter by lazy { RecommendedTvAdapter(::onClickRecommendationItem) }
+    private val reviewsAdapter by lazy { ReviewsAdapter(::onClickReviewItem) }
+    private val videoAdapter by lazy { VideoAdapter(::onClickVideoTrailerItem) }
     private var filmId = 0
 
     override fun getLayout(): Int {
@@ -47,7 +41,6 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
     }
 
     override fun onCreateViewSetup(savedInstanceState: Bundle?) {
-        (activity as DetailActivity).getDetailComponent().inject(this)
         arguments?.let {
             filmId = DetailTvFragmentArgs.fromBundle(it).filmId
         }
@@ -59,17 +52,11 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
 
     override fun onMain(savedInstanceState: Bundle?) {
         setupRecycler()
-        adapterOnCLick()
         subscribeUI()
         viewModel.getDetailTvRequest(filmId)
     }
 
     private fun setupRecycler(){
-        seasonAdapter = SeasonAdapter()
-        creditsAdapter = CreditsAdapter()
-        recommendedTvAdapter = RecommendedTvAdapter()
-        reviewsAdapter = ReviewsAdapter()
-        videoAdapter = VideoAdapter()
 
         binding.apply {
             dtSeasonList.layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
@@ -89,23 +76,23 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
 
     private fun subscribeUI(){
         viewModel.apply {
-            detailTvRes.observe(viewLifecycleOwner, Observer { res ->
+            detailTvRes.observe(viewLifecycleOwner, { res ->
                 when(res.status){
                     SUCCESS -> {
                         hideLoading()
                         binding.tv = res?.data
                         seasonAdapter.submitList(res?.data?.seasons)
                     }
-                    Result.Status.ERROR -> {
+                    ERROR -> {
                         showNetworkError(false){
                             viewModel.getDetailTvRequest(filmId)
                         }
                     }
-                    Result.Status.LOADING -> { showLoading() }
+                    LOADING -> { showLoading() }
                 }
             })
 
-            creditMovieRes.observe(viewLifecycleOwner, Observer { res ->
+            creditMovieRes.observe(viewLifecycleOwner, { res ->
                 when(res.status){
                     SUCCESS -> {
                         if(res.data?.credits.isNullOrEmpty()) binding.dtCastEmpty.visible()
@@ -116,7 +103,7 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
                 }
             })
 
-            recommendationTvRes.observe(viewLifecycleOwner, Observer { res ->
+            recommendationTvRes.observe(viewLifecycleOwner, { res ->
                 when(res.status){
                     SUCCESS -> {
                         if(res.data?.results.isNullOrEmpty()) binding.dtRecomEmpty.visible()
@@ -127,7 +114,7 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
                 }
             })
 
-            trailerTvRes.observe(viewLifecycleOwner, Observer { res ->
+            trailerTvRes.observe(viewLifecycleOwner, { res ->
                 when(res.status){
                     SUCCESS -> {
                         if(res.data?.videos.isNullOrEmpty()) binding.videoEmpty.visible()
@@ -138,53 +125,17 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
                 }
             })
 
-            reviewsTvRes.observe(viewLifecycleOwner, Observer { res ->
+            reviewsTvRes.observe(viewLifecycleOwner, { res ->
                 when(res.status){
                     SUCCESS -> {
-                        if(res.data?.reviewList.isNullOrEmpty()) binding.dtReviewEmpty.visible()
+                        if(res.data?.results.isNullOrEmpty()) binding.dtReviewEmpty.visible()
                         else binding.dtReviewEmpty.gone()
-                        reviewsAdapter.submitList(res.data?.reviewList)
+                        reviewsAdapter.submitList(res.data?.results)
                     }
                     else -> {}
                 }
             })
         }
-    }
-
-    private fun adapterOnCLick(){
-
-        creditsAdapter.setOnClickListener(object: OnAdapterListener<Credits>{
-            override fun onClick(view: View, item: Credits) {
-                val action = DetailTvFragmentDirections.actionDetailTvFragmentToPersonFragment2(item.id ?: 0)
-                Navigation.findNavController(view).navigate(action)
-            }
-        })
-
-        recommendedTvAdapter.setOnClickListener(object: OnAdapterListener<Tv>{
-            override fun onClick(view: View, item: Tv) {
-                val bundle = Bundle().apply {
-                    putInt("filmId", item.id ?: 0)
-                    putString("type", Constant.TV)
-                }
-                changeActivity<DetailActivity>(bundle)
-            }
-        })
-
-        reviewsAdapter.setOnClickListener(object: OnAdapterListener<Reviews>{
-            override fun onClick(view: View, item: Reviews) {
-                val uri: Uri = Uri.parse(item.url)
-                val intent = Intent(Intent.ACTION_VIEW, uri)
-                context?.startActivity(intent)
-            }
-        })
-
-        videoAdapter.setOnClickAdapter(object: OnAdapterListener<Videos>{
-            override fun onClick(view: View, item: Videos) {
-                val bundle = Bundle()
-                bundle.putString("key", item.key)
-                changeActivity<DetailActivity>(bundle)
-            }
-        })
     }
 
     private fun showLoading(){
@@ -201,4 +152,29 @@ class DetailTvFragment : BaseFragment<FragmentDetailTvBinding>() {
         }
     }
 
+    private fun onClickCreditItem(credit: Credits) {
+        val action = DetailTvFragmentDirections.actionDetailTvFragmentToPersonFragment2(credit.id ?: 0)
+        view.navigateFragment { Navigation.findNavController(it).navigate(action) }
+    }
+
+    private fun onClickRecommendationItem(tv: TvResponse) {
+        val bundle = Bundle().apply {
+            putInt("filmId", tv.id ?: 0)
+            putString("type", Constant.TV)
+        }
+        changeActivity<DetailActivity>(bundle)
+    }
+
+    private fun onClickVideoTrailerItem(video: Videos) {
+        val bundle = Bundle().apply {
+            putString("key", video.key)
+        }
+        changeActivity<YoutubeActivity>(bundle)
+    }
+
+    private fun onClickReviewItem(review: ReviewsResponse) {
+        val uri: Uri = Uri.parse(review.url)
+        val intent = Intent(Intent.ACTION_VIEW, uri)
+        requireContext().startActivity(intent)
+    }
 }
