@@ -1,45 +1,35 @@
 package com.themovie.ui.detail.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.themovie.R
+import com.aldebaran.data.network.ApiUrl
+import com.aldebaran.domain.entities.remote.MovieResponse
 import com.themovie.databinding.AdapterRecomendedBinding
-import com.themovie.helper.OnAdapterListener
 import com.themovie.helper.customview.PortraitView
-import com.themovie.model.db.Movies
-import com.themovie.restapi.ApiUrl
-import kotlinx.android.synthetic.main.adapter_recomended.view.*
 
-class RecommendedAdapter : ListAdapter<Movies, RecommendedAdapter.ViewHolder>(DIFF_CALLBACK) {
-
-    private lateinit var context: Context
-    private lateinit var onClickAdapterListener: OnAdapterListener<Movies>
+class RecommendedAdapter (
+    private val onItemClick: (MovieResponse) -> Unit
+) : ListAdapter<MovieResponse, RecommendedAdapter.ViewHolder>(DIFF_CALLBACK) {
 
     companion object{
-        val DIFF_CALLBACK: DiffUtil.ItemCallback<Movies> = object: DiffUtil.ItemCallback<Movies>(){
-            override fun areItemsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+        val DIFF_CALLBACK: DiffUtil.ItemCallback<MovieResponse> = object: DiffUtil.ItemCallback<MovieResponse>(){
+            override fun areItemsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.id == newItem.id
             }
 
-            override fun areContentsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+            override fun areContentsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.title == newItem.title && oldItem.backdropPath == newItem.backdropPath
             }
         }
     }
 
-    fun setOnClickListener(onClickAdapterListener: OnAdapterListener<Movies>){
-        this.onClickAdapterListener = onClickAdapterListener
-    }
-
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = AdapterRecomendedBinding
             .inflate(LayoutInflater.from(parent.context), parent, false)
-        context = parent.context
         return ViewHolder(view.root, view)
     }
 
@@ -48,7 +38,7 @@ class RecommendedAdapter : ListAdapter<Movies, RecommendedAdapter.ViewHolder>(DI
     }
 
     inner class ViewHolder(root: View, val binding: AdapterRecomendedBinding) : RecyclerView.ViewHolder(root){
-        fun bindItem(movies: Movies){
+        fun bindItem(movies: MovieResponse){
             val imgUrl = "${ApiUrl.IMG_POSTER}${movies.posterPath.toString()}"
             binding.recItem.apply {
                 setImage(imgUrl)
@@ -56,7 +46,7 @@ class RecommendedAdapter : ListAdapter<Movies, RecommendedAdapter.ViewHolder>(DI
                 setRating(movies.voteAverage.orEmpty())
                 setOnClickListener(object: PortraitView.OnClickListener{
                     override fun onClick() {
-                        onClickAdapterListener.onClick(itemView, movies)
+                        onItemClick.invoke(movies)
                     }
                 })
             }
