@@ -1,46 +1,35 @@
 package com.themovie.ui.discover
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.View
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import androidx.navigation.Navigation
-import androidx.paging.PagedList
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.aldebaran.base.BaseFragment
+import com.aldebaran.domain.entities.remote.MovieResponse
+import com.aldebaran.utils.changeActivity
+import com.aldebaran.utils.initLinearRecycler
 
 import com.themovie.R
-import com.themovie.base.BaseFragment
 import com.themovie.databinding.FragmentMoviesBinding
-import com.themovie.di.main.MainViewModelFactory
 import com.themovie.helper.Constant
-import com.themovie.helper.OnAdapterListener
-import com.themovie.helper.changeActivity
-import com.themovie.model.db.Movies
 import com.themovie.ui.detail.DetailActivity
 import com.themovie.ui.discover.adapter.MovieAdapter
-import com.themovie.ui.main.MainActivity
 import com.themovie.ui.search.SuggestActivity
-import kotlinx.android.synthetic.main.fragment_movies.*
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
-/**
- * A simple [Fragment] subclass.
- */
+@AndroidEntryPoint
 class MoviesFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefreshLayout.OnRefreshListener {
 
-    @Inject lateinit var factory: MainViewModelFactory
-    private val viewModel by viewModels<MovieViewModel>{ factory }
-    private lateinit var mAdapter: MovieAdapter
+    private val viewModel by viewModels<MovieViewModel>()
+    private val mAdapter by lazy { MovieAdapter(::onMovieItemClick, ::onLoadMoreRetry) }
 
     override fun getLayout(): Int {
         return R.layout.fragment_movies
     }
 
     override fun onCreateViewSetup(savedInstanceState: Bundle?) {
-        (activity as MainActivity).getMainComponent()?.inject(this)
         binding.apply {
             vm = viewModel
             lifecycleOwner = this@MoviesFragment
@@ -68,17 +57,17 @@ class MoviesFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefreshLayout
     }
 
     private fun setupUIComponent(){
-        swipe.setOnRefreshListener(this)
+        binding.swipe.setOnRefreshListener(this)
         binding.header.apply {
             setLogoVisibility(View.GONE)
             setBackButtonVisibility(View.VISIBLE)
             setTitleText(resources.getString(R.string.home_title_5))
-            setBackButtonOnClickListener(View.OnClickListener {
+            setBackButtonOnClickListener {
                 val action = MoviesFragmentDirections.actionMoviesFragmentToHomeFragment()
                 Navigation.findNavController(it).navigate(action)
-            })
+            }
 
-            setSearchButtonOnClickListener(View.OnClickListener { changeActivity<SuggestActivity>() })
+            setSearchButtonOnClickListener { changeActivity<SuggestActivity>() }
         }
 
 
@@ -92,40 +81,30 @@ class MoviesFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefreshLayout
     }
 
     private fun recyclerViewSetup(){
-        mAdapter = MovieAdapter()
-        movie_rec.apply {
-            layoutManager = LinearLayoutManager(context)
-            adapter = mAdapter
-        }
-
-        mAdapter.setOnClickAdapter(object: OnAdapterListener<Movies>{
-            override fun onClick(view: View, item: Movies) {
-                val bundle = Bundle().apply {
-                    putInt("filmId", item.id ?: 0)
-                    putString("type", Constant.MOVIE)
-                }
-                changeActivity<DetailActivity>(bundle)
-            }
-        })
-
-        mAdapter.setOnErrorClickListener(object: MovieAdapter.OnErrorClickListener{
-            override fun onClick(view: View?) {
-                viewModel.retry()
-            }
-        })
+        binding.movieRec.initLinearRecycler(requireContext())
+        binding.movieRec.adapter = mAdapter
     }
 
     private fun getDiscoverMovies(){
         viewModel.apply {
-            getMovieLiveData().observe(this@MoviesFragment,
-                Observer<PagedList<Movies>>{
+            getMovieLiveData().observe(this@MoviesFragment, {
                     mAdapter.submitList(it)
-                    swipe.isRefreshing = false
-                })
+                    binding.swipe.isRefreshing = false
+            })
 
-            getLoadState().observe(this@MoviesFragment,
-                Observer{ mAdapter.setLoadState(it) })
+            getLoadState().observe(this@MoviesFragment, { mAdapter.setLoadState(it) })
         }
     }
 
+    private fun onMovieItemClick(movie: MovieResponse) {
+        val bundle = Bundle().apply {
+            putInt("filmId", movie.id ?: 0)
+            putString("type", Constant.MOVIE)
+        }
+        changeActivity<DetailActivity>(bundle)
+    }
+
+    private fun onLoadMoreRetry() {
+        viewModel.retry()
+    }
 }
