@@ -1,34 +1,36 @@
 package com.themovie.ui.discover.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.paging.PagedListAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.aldebaran.domain.Result.*
+import com.aldebaran.domain.Result.Status.*
+import com.aldebaran.domain.entities.remote.MovieResponse
 import com.themovie.databinding.AdapterLoadingBinding
 import com.themovie.databinding.AdapterMoviesBinding
-import com.themovie.helper.OnAdapterListener
-import com.themovie.model.db.Movies
-import com.themovie.restapi.Result
 
-class MovieAdapter: PagedListAdapter<Movies, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
+class MovieAdapter (
+    private val onItemClick: (MovieResponse) -> Unit,
+    private val onItemErrorClick: () -> Unit
+): PagedListAdapter<MovieResponse, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
 
-    private val DATA_VIEW = 1
-    private val LOADING_VIEW = 2
-    private var loadState: Result.Status? = null
-    private lateinit var context: Context
-    private lateinit var onErrorClickListener: OnErrorClickListener
-    private lateinit var onClickAdapterListener: OnAdapterListener<Movies>
+
+    private var loadState: Status? = null
 
     companion object{
-        val DIFF_CALLBACK: DiffUtil.ItemCallback<Movies> = object: DiffUtil.ItemCallback<Movies>(){
-            override fun areItemsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+
+        private const val DATA_VIEW = 1
+        private const val LOADING_VIEW = 2
+
+        val DIFF_CALLBACK: DiffUtil.ItemCallback<MovieResponse> = object: DiffUtil.ItemCallback<MovieResponse>(){
+            override fun areItemsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.id == newItem.id
             }
 
-            override fun areContentsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+            override fun areContentsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.title == newItem.title &&
                         oldItem.posterPath == newItem.posterPath &&
                          oldItem.backdropPath == newItem.backdropPath
@@ -37,8 +39,7 @@ class MovieAdapter: PagedListAdapter<Movies, RecyclerView.ViewHolder>(DIFF_CALLB
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        context = parent.context
-        val inflater = LayoutInflater.from(context)
+        val inflater = LayoutInflater.from(parent.context)
         return if(viewType == DATA_VIEW){
             val view = AdapterMoviesBinding
                 .inflate(inflater, parent, false)
@@ -55,7 +56,7 @@ class MovieAdapter: PagedListAdapter<Movies, RecyclerView.ViewHolder>(DIFF_CALLB
             holder.binding.movie = getItem(position)
             holder.binding.vh = holder
         }else if(holder is LoadingViewHolder){
-            holder.bindView(loadState, onErrorClickListener)
+            holder.bindView(loadState, onItemErrorClick)
         }
     }
 
@@ -71,10 +72,10 @@ class MovieAdapter: PagedListAdapter<Movies, RecyclerView.ViewHolder>(DIFF_CALLB
     }
 
     private fun hasFooter(): Boolean {
-        return loadState != null && loadState != Result.Status.SUCCESS
+        return loadState != null && loadState != SUCCESS
     }
 
-    fun setLoadState(loadState: Result.Status){
+    fun setLoadState(loadState: Status){
         val previousState = this.loadState
         val previousExtraRow = hasFooter()
         this.loadState = loadState
@@ -87,22 +88,10 @@ class MovieAdapter: PagedListAdapter<Movies, RecyclerView.ViewHolder>(DIFF_CALLB
         }
     }
 
-    fun setOnErrorClickListener(onErrorClickListener: OnErrorClickListener){
-        this.onErrorClickListener = onErrorClickListener
-    }
-
-    fun setOnClickAdapter(onClickAdapterListener: OnAdapterListener<Movies>){
-        this.onClickAdapterListener = onClickAdapterListener
-    }
-
-    interface OnErrorClickListener {
-        fun onClick(view: View?)
-    }
-
     inner class MovieViewHolder(root: View, val binding: AdapterMoviesBinding) : RecyclerView.ViewHolder(root){
 
-        fun onMovieClick(view: View, movies: Movies){
-            onClickAdapterListener.onClick(view, movies)
+        fun onMovieClick(movies: MovieResponse){
+            onItemClick.invoke(movies)
         }
     }
 }
