@@ -1,34 +1,35 @@
 package com.themovie.ui.discover.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.paging.PagedListAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.aldebaran.domain.Result.*
+import com.aldebaran.domain.Result.Status.SUCCESS
+import com.aldebaran.domain.entities.remote.TvResponse
 import com.themovie.databinding.AdapterLoadingBinding
 import com.themovie.databinding.AdapterTvBinding
-import com.themovie.helper.OnAdapterListener
-import com.themovie.model.db.Tv
-import com.themovie.restapi.Result
 
-class TvAdapter: PagedListAdapter<Tv, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
+class TvAdapter (
+    private val onItemClick: (TvResponse) -> Unit,
+    private val onItemErrorClick: () -> Unit
+): PagedListAdapter<TvResponse, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
 
-    private val DATA_VIEW = 1
-    private val LOADING_VIEW = 2
-    private var loadState: Result.Status? = null
-    private lateinit var context: Context
-    private lateinit var onErrorClickListener: OnErrorClickListener
-    private lateinit var onClickAdapterListener: OnAdapterListener<Tv>
+    private var loadState: Status? = null
 
     companion object{
-        val DIFF_CALLBACK: DiffUtil.ItemCallback<Tv> = object: DiffUtil.ItemCallback<Tv>(){
-            override fun areItemsTheSame(oldItem: Tv, newItem: Tv): Boolean {
+
+        private const val DATA_VIEW = 1
+        private const val LOADING_VIEW = 2
+
+        val DIFF_CALLBACK: DiffUtil.ItemCallback<TvResponse> = object: DiffUtil.ItemCallback<TvResponse>(){
+            override fun areItemsTheSame(oldItem: TvResponse, newItem: TvResponse): Boolean {
                 return oldItem.id == newItem.id
             }
 
-            override fun areContentsTheSame(oldItem: Tv, newItem: Tv): Boolean {
+            override fun areContentsTheSame(oldItem: TvResponse, newItem: TvResponse): Boolean {
                 return oldItem.name == newItem.name &&
                         oldItem.posterPath == newItem.posterPath &&
                          oldItem.backdropPath == newItem.backdropPath
@@ -37,8 +38,7 @@ class TvAdapter: PagedListAdapter<Tv, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        context = parent.context
-        val inflater = LayoutInflater.from(context)
+        val inflater = LayoutInflater.from(parent.context)
         return if(viewType == DATA_VIEW){
             val view = AdapterTvBinding
                 .inflate(inflater, parent, false)
@@ -55,7 +55,7 @@ class TvAdapter: PagedListAdapter<Tv, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
             holder.binding.tv = getItem(position)
             holder.binding.vh = holder
         } else if(holder is LoadingViewHolder){
-            holder.bindView(loadState, onErrorClickListener)
+            holder.bindView(loadState, onItemErrorClick)
         }
     }
 
@@ -71,10 +71,10 @@ class TvAdapter: PagedListAdapter<Tv, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
     }
 
     private fun hasFooter(): Boolean {
-        return loadState != null && loadState != Result.Status.SUCCESS
+        return loadState != null && loadState != SUCCESS
     }
 
-    fun setLoadState(loadState: Result.Status){
+    fun setLoadState(loadState: Status){
         val previousState = this.loadState
         val previousExtraRow = hasFooter()
         this.loadState = loadState
@@ -87,21 +87,13 @@ class TvAdapter: PagedListAdapter<Tv, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
         }
     }
 
-    fun setOnErrorClickListener(onErrorClickListener: OnErrorClickListener){
-        this.onErrorClickListener = onErrorClickListener
-    }
-
-    fun setOnClickAdapter(onClickAdapterListener: OnAdapterListener<Tv>){
-        this.onClickAdapterListener = onClickAdapterListener
-    }
-
     interface OnErrorClickListener {
         fun onClick(view: View?)
     }
 
     inner class TvViewHolder(root: View, val binding: AdapterTvBinding) : RecyclerView.ViewHolder(root) {
-        fun onTvAdapterClick(view: View, tv: Tv){
-            onClickAdapterListener.onClick(view, tv)
+        fun onTvAdapterClick(tv: TvResponse){
+            onItemClick.invoke(tv)
         }
     }
 }
