@@ -1,34 +1,33 @@
 package com.themovie.ui.discover.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.paging.PagedListAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
+import com.aldebaran.domain.Result.*
+import com.aldebaran.domain.Result.Status.*
+import com.aldebaran.domain.entities.remote.MovieResponse
 import com.themovie.databinding.AdapterLoadingBinding
 import com.themovie.databinding.AdapterUpcomingBinding
-import com.themovie.helper.OnAdapterListener
-import com.themovie.model.db.Upcoming
-import com.themovie.restapi.Result
 
-class UpcomingAdapter: PagedListAdapter<Upcoming, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
+class UpcomingAdapter (
+    private val onItemClick: (MovieResponse) -> Unit,
+    private val onItemErrorClick: () -> Unit
+): PagedListAdapter<MovieResponse, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
 
-    private val DATA_VIEW = 1
-    private val LOADING_VIEW = 2
-    private var loadState: Result.Status? = null
-    private lateinit var context: Context
-    private lateinit var onErrorClickListener: OnErrorClickListener
-    private lateinit var onClickAdapterListener: OnAdapterListener<Upcoming>
+    private var loadState: Status? = null
 
     companion object{
-        val DIFF_CALLBACK: DiffUtil.ItemCallback<Upcoming> = object: DiffUtil.ItemCallback<Upcoming>(){
-            override fun areItemsTheSame(oldItem: Upcoming, newItem: Upcoming): Boolean {
+        private const val DATA_VIEW = 1
+        private const val LOADING_VIEW = 2
+        val DIFF_CALLBACK: DiffUtil.ItemCallback<MovieResponse> = object: DiffUtil.ItemCallback<MovieResponse>(){
+            override fun areItemsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.id == newItem.id
             }
 
-            override fun areContentsTheSame(oldItem: Upcoming, newItem: Upcoming): Boolean {
+            override fun areContentsTheSame(oldItem: MovieResponse, newItem: MovieResponse): Boolean {
                 return oldItem.title == newItem.title &&
                         oldItem.posterPath == newItem.posterPath &&
                         oldItem.backdropPath == newItem.backdropPath
@@ -37,8 +36,7 @@ class UpcomingAdapter: PagedListAdapter<Upcoming, RecyclerView.ViewHolder>(DIFF_
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        context = parent.context
-        val inflater = LayoutInflater.from(context)
+        val inflater = LayoutInflater.from(parent.context)
         return if(viewType == DATA_VIEW){
             val view = AdapterUpcomingBinding
                 .inflate(inflater, parent, false)
@@ -55,7 +53,7 @@ class UpcomingAdapter: PagedListAdapter<Upcoming, RecyclerView.ViewHolder>(DIFF_
             holder.binding.upcoming = getItem(position)
             holder.binding.vh = holder
         }else if(holder is LoadingViewHolder){
-            holder.bindView(loadState, onErrorClickListener)
+            holder.bindView(loadState, onItemErrorClick)
         }
     }
 
@@ -71,10 +69,10 @@ class UpcomingAdapter: PagedListAdapter<Upcoming, RecyclerView.ViewHolder>(DIFF_
     }
 
     private fun hasFooter(): Boolean {
-        return loadState != null && loadState != Result.Status.SUCCESS
+        return loadState != null && loadState != SUCCESS
     }
 
-    fun setLoadState(loadState: Result.Status){
+    fun setLoadState(loadState: Status){
         val previousState = this.loadState
         val previousExtraRow = hasFooter()
         this.loadState = loadState
@@ -87,21 +85,9 @@ class UpcomingAdapter: PagedListAdapter<Upcoming, RecyclerView.ViewHolder>(DIFF_
         }
     }
 
-    fun setOnErrorClickListener(onErrorClickListener: OnErrorClickListener){
-        this.onErrorClickListener = onErrorClickListener
-    }
-
-    fun setOnClickAdapter(onClickAdapterListener: OnAdapterListener<Upcoming>){
-        this.onClickAdapterListener = onClickAdapterListener
-    }
-
-    interface OnErrorClickListener {
-        fun onClick(view: View?)
-    }
-
     inner class UpcomingViewHolder(root: View, val binding: AdapterUpcomingBinding) : RecyclerView.ViewHolder(root){
-        fun onUpcomingClick(view: View, data: Upcoming){
-            onClickAdapterListener.onClick(view, data)
+        fun onUpcomingClick(data: MovieResponse){
+            onItemClick.invoke(data)
         }
     }
 }
