@@ -2,37 +2,28 @@ package com.themovie.ui.genres
 
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.View
 import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import androidx.navigation.Navigation
-import androidx.paging.PagedList
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.aldebaran.base.BaseFragment
+import com.aldebaran.domain.entities.remote.MovieResponse
+import com.aldebaran.utils.changeActivity
+import com.aldebaran.utils.initLinearRecycler
 import com.themovie.R
-import com.themovie.base.BaseFragment
 import com.themovie.databinding.FragmentMoviesBinding
-import com.themovie.di.main.MainViewModelFactory
 import com.themovie.helper.Constant
-import com.themovie.helper.OnAdapterListener
-import com.themovie.helper.changeActivity
-import com.themovie.model.db.Movies
 import com.themovie.ui.detail.DetailActivity
 import com.themovie.ui.discover.MovieViewModel
 import com.themovie.ui.discover.adapter.MovieAdapter
-import com.themovie.ui.main.MainActivity
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
-/**
- * A simple [Fragment] subclass.
- */
+@AndroidEntryPoint
 class MovieWithGenreFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefreshLayout.OnRefreshListener {
 
-    @Inject lateinit var factory: MainViewModelFactory
-    private val viewModel by viewModels<MovieViewModel> { factory }
-    private lateinit var mAdapter: MovieAdapter
+    private val viewModel by viewModels<MovieViewModel>()
+    private val mAdapter by lazy { MovieAdapter(::onMovieItemClick, ::onLoadMoreRetry) }
     private var destinationBackPress = ""
     private var title = ""
     private var genreId = ""
@@ -42,7 +33,6 @@ class MovieWithGenreFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefre
     }
 
     override fun onCreateViewSetup(savedInstanceState: Bundle?) {
-        (activity as MainActivity).getMainComponent()?.inject(this)
         arguments?.let {
             genreId = MovieWithGenreFragmentArgs.fromBundle(it).genreId.toString()
             destinationBackPress = MovieWithGenreFragmentArgs.fromBundle(it).from
@@ -65,9 +55,9 @@ class MovieWithGenreFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefre
             setBackButtonVisibility(View.VISIBLE)
             setSearchVisibility(View.GONE)
             setTitleText("Genres: $title")
-            setBackButtonOnClickListener(View.OnClickListener {
+            setBackButtonOnClickListener {
                 Navigation.findNavController(it).navigate(action)
-            })
+            }
         }
         binding.swipe.setOnRefreshListener(this)
         viewModel.resetMovieWithGenre(genreId)
@@ -95,41 +85,34 @@ class MovieWithGenreFragment : BaseFragment<FragmentMoviesBinding>(), SwipeRefre
     }
 
     private fun recyclerViewSetup(){
-        mAdapter = MovieAdapter()
         binding.movieRec.apply {
-            layoutManager = LinearLayoutManager(context)
+            initLinearRecycler(requireContext())
             adapter = mAdapter
         }
-
-        mAdapter.setOnClickAdapter(object: OnAdapterListener<Movies>{
-            override fun onClick(view: View, item: Movies) {
-                val bundle = Bundle().apply {
-                    putInt("filmId", item.id ?: 0)
-                    putString("type", Constant.MOVIE)
-                }
-                changeActivity<DetailActivity>(bundle)
-            }
-        })
-
-        mAdapter.setOnErrorClickListener(object: MovieAdapter.OnErrorClickListener{
-            override fun onClick(view: View?) {
-                viewModel.retry()
-            }
-        })
     }
 
     private fun getDiscoverMovies(){
         viewModel.apply {
-            getMovieLiveData().observe(this@MovieWithGenreFragment,
-                Observer<PagedList<Movies>>{
+            getMovieLiveData().observe(this@MovieWithGenreFragment, {
                     mAdapter.submitList(it)
                     binding.swipe.isRefreshing = false
-                })
+            })
 
-            getLoadState().observe(this@MovieWithGenreFragment,
-                Observer{
+            getLoadState().observe(this@MovieWithGenreFragment, {
                     mAdapter.setLoadState(it)
-                })
+            })
         }
+    }
+
+    private fun onMovieItemClick(movie: MovieResponse) {
+        val bundle = Bundle().apply {
+            putInt("filmId", movie.id ?: 0)
+            putString("type", Constant.MOVIE)
+        }
+        changeActivity<DetailActivity>(bundle)
+    }
+
+    private fun onLoadMoreRetry() {
+        viewModel.retry()
     }
 }
