@@ -1,46 +1,37 @@
 package com.themovie.ui.main.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.aldebaran.data.network.ApiUrl
+import com.aldebaran.domain.entities.local.MovieEntity
 import com.themovie.databinding.AdapterPortraitMovieBinding
-import com.themovie.helper.OnAdapterListener
 import com.themovie.helper.customview.PortraitView
-import com.themovie.model.db.Movies
-import com.themovie.restapi.ApiUrl
 
-class DiscoverMovieAdapter : ListAdapter<Movies, DiscoverMovieAdapter.ViewHolder>(DIFF_CALLBACK) {
-
-    private lateinit var context: Context
-    private lateinit var listener: OnAdapterListener<Movies>
+class DiscoverMovieAdapter (
+    private val onItemClick: (MovieEntity) -> Unit
+) : ListAdapter<MovieEntity, DiscoverMovieAdapter.ViewHolder>(DIFF_CALLBACK) {
 
     companion object{
-        val DIFF_CALLBACK: DiffUtil.ItemCallback<Movies> = object: DiffUtil.ItemCallback<Movies>(){
-            override fun areItemsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+        val DIFF_CALLBACK: DiffUtil.ItemCallback<MovieEntity> = object: DiffUtil.ItemCallback<MovieEntity>(){
+            override fun areItemsTheSame(oldItem: MovieEntity, newItem: MovieEntity): Boolean {
                 return oldItem.id == newItem.id
             }
 
-            override fun areContentsTheSame(oldItem: Movies, newItem: Movies): Boolean {
+            override fun areContentsTheSame(oldItem: MovieEntity, newItem: MovieEntity): Boolean {
                 return oldItem.title == newItem.title && oldItem.posterPath == newItem.posterPath
             }
         }
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        context = parent.context
-
         val view = AdapterPortraitMovieBinding
-            .inflate(LayoutInflater.from(context), parent, false)
+            .inflate(LayoutInflater.from(parent.context), parent, false)
 
         return ViewHolder(view.root, view)
-    }
-
-    fun setOnClickListener(listener: OnAdapterListener<Movies>){
-        this.listener = listener
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
@@ -48,15 +39,15 @@ class DiscoverMovieAdapter : ListAdapter<Movies, DiscoverMovieAdapter.ViewHolder
     }
 
     inner class ViewHolder(root: View, val binding: AdapterPortraitMovieBinding) : RecyclerView.ViewHolder(root){
-        fun bindItem(movies: Movies){
+        fun bindItem(movies: MovieEntity){
             val imgPoster = "${ApiUrl.IMG_POSTER}${movies.posterPath}"
             binding.movieItem.apply {
                 setImage(imgPoster)
                 setTitle(movies.title.orEmpty())
                 setRating(movies.voteAverage.orEmpty())
-                setOnClickListener(object: PortraitView.OnClickListener{
+                setOnClickListener(object : PortraitView.OnClickListener{
                     override fun onClick() {
-                        listener.onClick(itemView, movies)
+                        onItemClick.invoke(movies)
                     }
                 })
             }
