@@ -2,13 +2,14 @@ package com.themovie.ui.search
 
 import android.os.Bundle
 import android.view.View
+import com.aldebaran.base.BaseActivity
+import com.aldebaran.utils.changeActivity
 import com.themovie.R
-import com.themovie.base.BaseActivity
 import com.themovie.databinding.ActivitySearchBinding
-import com.themovie.di.search.SearchComponent
 import com.themovie.helper.ViewPagerFragment
-import com.themovie.helper.changeActivity
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class SearchActivity : BaseActivity<ActivitySearchBinding>() {
 
     private var query: String? = ""
@@ -55,14 +56,5 @@ class SearchActivity : BaseActivity<ActivitySearchBinding>() {
             hSearch.apply { isFocusable = false; setText(query) }
             pagerAdapter.notifyDataSetChanged()
         }
-    }
-
-    fun getComponent(): SearchComponent? {
-        return getApp().getSearchComponent()
-    }
-
-    override fun onDestroy() {
-        getApp().releaseSearchComponent()
-        super.onDestroy()
     }
 }
