@@ -2,35 +2,27 @@ package com.themovie.ui.search
 
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
-import androidx.paging.PagedList
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+import com.aldebaran.base.BaseFragment
+import com.aldebaran.domain.entities.remote.TvResponse
+import com.aldebaran.utils.changeActivity
+import com.aldebaran.utils.initLinearRecycler
 
 import com.themovie.R
-import com.themovie.base.BaseFragment
 import com.themovie.databinding.FragmentSearchResultBinding
-import com.themovie.di.search.SearchViewModelFactory
 import com.themovie.helper.Constant
-import com.themovie.helper.OnAdapterListener
-import com.themovie.helper.changeActivity
-import com.themovie.model.db.Tv
 import com.themovie.ui.detail.DetailActivity
 import com.themovie.ui.discover.adapter.TvAdapter
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
-/**
- * A simple [Fragment] subclass.
- */
+@AndroidEntryPoint
 class SearchTvFragment : BaseFragment<FragmentSearchResultBinding>(), SwipeRefreshLayout.OnRefreshListener {
 
-    @Inject lateinit var factory: SearchViewModelFactory
     private var query: String? = ""
-    private val viewModel by viewModels<SearchTvViewModel> { factory }
-    private lateinit var mAdapter: TvAdapter
+    private val viewModel by viewModels<SearchTvViewModel>()
+    private val mAdapter by lazy { TvAdapter(::onTvShowItemClick, ::onLoadMoreRetry) }
 
     override fun getLayout(): Int {
         return R.layout.fragment_search_result
@@ -38,8 +30,6 @@ class SearchTvFragment : BaseFragment<FragmentSearchResultBinding>(), SwipeRefre
 
     override fun onCreateViewSetup(savedInstanceState: Bundle?) {
         binding.lifecycleOwner = this
-        (activity as SearchActivity).getComponent()?.inject(this)
-
         query = getBundle()?.getString("query")
         SearchTvViewModel.query = query.orEmpty()
 
@@ -56,35 +46,16 @@ class SearchTvFragment : BaseFragment<FragmentSearchResultBinding>(), SwipeRefre
     }
 
     private fun setupRecyclerView(){
-        mAdapter = TvAdapter()
-        mAdapter.apply {
-            setOnClickAdapter(object: OnAdapterListener<Tv> {
-                override fun onClick(view: View, item: Tv) {
-                    val bundle = Bundle().apply {
-                        putInt("filmId", item.id ?: 0)
-                        putString("type", Constant.TV)
-                    }
-                    changeActivity<DetailActivity>(bundle)
-                }
-            })
-
-            setOnErrorClickListener(object: TvAdapter.OnErrorClickListener{
-                override fun onClick(view: View?) {
-                    viewModel.retry()
-                }
-            })
-        }
 
         binding.recyclerView.apply {
-            layoutManager = LinearLayoutManager(context)
+            initLinearRecycler(requireContext())
             adapter = mAdapter
         }
     }
 
     private fun getSearchResult(){
         viewModel.apply {
-            getSearchTvResult().observe(this@SearchTvFragment,
-                Observer<PagedList<Tv>>{
+            getSearchTvResult().observe(this@SearchTvFragment, {
                     mAdapter.submitList(it)
                     binding.swipe.isRefreshing = false
                 })
@@ -92,5 +63,17 @@ class SearchTvFragment : BaseFragment<FragmentSearchResultBinding>(), SwipeRefre
             getLoadState().observe(this@SearchTvFragment,
                 Observer { mAdapter.setLoadState(it) })
         }
+    }
+
+    private fun onTvShowItemClick(tv: TvResponse) {
+        val bundle = Bundle().apply {
+            putInt("filmId", tv.id ?: 0)
+            putString("type", Constant.TV)
+        }
+        changeActivity<DetailActivity>(bundle)
+    }
+
+    private fun onLoadMoreRetry() {
+        viewModel.retry()
     }
 }
