@@ -2,39 +2,34 @@ package com.themovie.ui.search
 
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.View
+import androidx.core.view.isVisible
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import androidx.recyclerview.widget.LinearLayoutManager
+import com.aldebaran.base.BaseFragment
+import com.aldebaran.domain.entities.remote.TvResponse
 
 import com.themovie.R
-import com.themovie.base.BaseFragment
 import com.themovie.databinding.FragmentSuggestBinding
-import com.themovie.di.suggest.SuggestViewModelFactory
+import com.aldebaran.domain.Result.Status.*
+import com.aldebaran.utils.changeActivity
+import com.aldebaran.utils.gone
+import com.aldebaran.utils.initLinearRecycler
+import com.aldebaran.utils.visible
 import com.themovie.helper.*
-import com.themovie.model.db.Tv
-import com.themovie.model.online.discovertv.TvResponse
-import com.themovie.restapi.Result
 import com.themovie.ui.detail.DetailActivity
 import com.themovie.ui.search.adapter.SuggestTvAdapter
-import javax.inject.Inject
+import dagger.hilt.android.AndroidEntryPoint
 
-/**
- * A simple [Fragment] subclass.
- */
+@AndroidEntryPoint
 class SuggestTvFragment : BaseFragment<FragmentSuggestBinding>(), SuggestActivity.TvSearchFragmentListener {
 
-    @Inject lateinit var factory: SuggestViewModelFactory
-    private val viewModel by viewModels<SuggestTvViewModel> { factory }
-    private lateinit var mAdapter: SuggestTvAdapter
+    private val viewModel by viewModels<SuggestTvViewModel>()
+    private val mAdapter by lazy { SuggestTvAdapter(::onTvShowItemClick) }
 
     override fun getLayout(): Int {
         return R.layout.fragment_suggest
     }
 
     override fun onCreateViewSetup(savedInstanceState: Bundle?) {
-        (activity as SuggestActivity).getComponent()?.inject(this)
         binding.lifecycleOwner = this
     }
 
@@ -45,35 +40,22 @@ class SuggestTvFragment : BaseFragment<FragmentSuggestBinding>(), SuggestActivit
     }
 
     private fun setupRecyclerView(){
-        mAdapter = SuggestTvAdapter()
-        mAdapter.setAdapterListener(object: OnAdapterListener<Tv> {
-            override fun onClick(view: View, item: Tv) {
-                val bundle = Bundle().apply {
-                    putInt("filmId", item.id ?: 0)
-                    putString("type", Constant.TV)
-                }
-                changeActivity<DetailActivity>(bundle)
-                activity?.finish()
-            }
-        })
-
         binding.recyclerView.apply {
-            layoutManager = LinearLayoutManager(context)
+            initLinearRecycler(requireContext())
             adapter = mAdapter
         }
     }
 
     private fun observeSuggestData(){
-        viewModel.getSuggestTv().observe(this,
-            Observer<Result<TvResponse>> {
+        viewModel.getSuggestTv().observe(this, {
                 when(it.status){
-                    Result.Status.LOADING -> {setLog("Loading")}
-                    Result.Status.SUCCESS -> {
-                        if(binding.recyclerView.visibility == View.GONE)
+                    LOADING -> {setLog("Loading")}
+                    SUCCESS -> {
+                        if(!binding.recyclerView.isVisible)
                             binding.recyclerView.visible()
                         mAdapter.submitList(it.data?.results)
                     }
-                    Result.Status.ERROR -> {
+                    ERROR -> {
                         binding.recyclerView.gone()
                     }
                 }
@@ -82,6 +64,14 @@ class SuggestTvFragment : BaseFragment<FragmentSuggestBinding>(), SuggestActivit
 
     override fun textChange(text: String) {
         viewModel.fetchSuggestTv(text)
+    }
 
+    private fun onTvShowItemClick(tv: TvResponse) {
+        val bundle = Bundle().apply {
+            putInt("filmId", tv.id ?: 0)
+            putString("type", Constant.TV)
+        }
+        changeActivity<DetailActivity>(bundle)
+        activity?.finish()
     }
 }
