@@ -1,12 +1,12 @@
 package com.themovie.ui.youtube
 
 import android.os.Bundle
+import com.aldebaran.data.network.ApiUrl
 import com.google.android.youtube.player.YouTubeBaseActivity
 import com.google.android.youtube.player.YouTubeInitializationResult
 import com.google.android.youtube.player.YouTubePlayer
 import com.themovie.BuildConfig
 import com.themovie.R
-import com.themovie.restapi.ApiUrl
 import kotlinx.android.synthetic.main.activity_youtube.*
 
 class YoutubeActivity : YouTubeBaseActivity(), YouTubePlayer.OnInitializedListener {
@@ -15,7 +15,7 @@ class YoutubeActivity : YouTubeBaseActivity(), YouTubePlayer.OnInitializedListen
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_youtube)
 
-        yt_player.initialize(BuildConfig.YT, this)
+        yt_player.initialize(ApiUrl.YOUTUBE_KEY, this)
     }
 
     override fun onInitializationSuccess(provider: YouTubePlayer.Provider?, youTubePlayer: YouTubePlayer?, p2: Boolean) {
